[OMNI-5538] For fixing the formatting of loyalty points earned while creating order from admin

+++ src/Omni/Controller/Cart/RedeemPoints.php
@@ -99,7 +99,7 @@ class RedeemPoints extends \Magento\Checkout\Controller\Cart
             return $this->_goBack();
         }
 
-        $loyaltyPoints = (int)$loyaltyPoints;
+        $loyaltyPoints = (float)$loyaltyPoints;
         try {
             $cartQuote          = $this->cart->getQuote();
             $itemsCount         = $cartQuote->getItemsCount();
